Fix pre-existing bug in kNearestNeighEuc
Once a nearest neighbour had bene identified, its distance from the query point was increased to prevent it from being re-selected.
However, it was increased to the maximum value of the distance across all points in the set S; if two or more points had an equal distance,
then the same point could be selected multiple times and the k-nearest neighbours returned would not be distinct.