Represent priors by a named tuple
Previously, a prior was specified by a list of length three, where
the first element was an integer representing the distribution,
and the remaining two elements represent parameters.

This commit replaces the list with a NamedTuple. As a future step,
it would be a good idea to replace it by a full class, and then
sampleTheParameterFromPrior() could call a .sample() method rather
than containing an if-statement (and similarly for
 compute_particle_weights()).